tests cleanup
