+++ tests/DevToolsRouteTest.php
@@ -1,10 +1,6 @@
 <?php
 
-namespace Dedoc\Scramble\Tests;
-
-use PHPUnit\Framework\Attributes\Test;
-
-class DevToolsRouteTest extends TestCase
+trait EnablesDevToolsRoute
 {
     public function getEnvironmentSetUp($app)
     {
@@ -13,14 +9,14 @@ class DevToolsRouteTest extends TestCase
         $app['config']->set('scramble.dev_tools', true);
         $app['config']->set('scramble.middleware', []);
     }
-
-    #[Test]
-    public function it_serves_the_dev_tools_bundle(): void
-    {
-        $response = $this->get('/_scramble/dev-tools/devtools.js')
-            ->assertOk()
-            ->assertHeader('content-type', 'text/javascript; charset=UTF-8');
-
-        $this->assertStringContainsString('no-store', (string) $response->headers->get('cache-control'));
-    }
 }
+
+uses(EnablesDevToolsRoute::class);
+
+it('serves the dev tools bundle', function () {
+    $response = $this->get('/_scramble/dev-tools/devtools.js')
+        ->assertOk()
+        ->assertHeader('content-type', 'text/javascript; charset=UTF-8');
+
+    expect($response->headers->get('cache-control'))->toContain('no-store');
+});
